# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.HistoryEntry:
    properties:
      jenis_status_mahasiswa_key:
        type: integer
      nama_pa:
        type: string
      periode:
        type: string
      semester:
        type: integer
      status:
        type: string
      tahun:
        type: integer
    type: object
...
  models.KritikSaran:
    properties:
      _id:
        type: integer
      nama:
        type: string
      nip:
        type: string
      periode:
        type: string
      saran:
        items:
          type: string
        type: array
      semester:
        type: string
      tahun:
        type: string
      unit:
        $ref: '#/definitions/models.Unit'
    type: object
  models.ListDetailResponse:
    properties:
      datas: {}
      messagae:
        type: string
    type: object
  models.ListResponse:
    properties:
      datas: {}
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      status:
        example: success
        type: string
    type: object
  models.Mahasiswa:
    properties:
      agama_key:
        type: integer
      alamat:
        type: string
      created_at:
        type: string
      dosen_pa_key:
        type: string
      email:
        type: string
      email_sso:
        type: string
      email2:
        type: string
      expiry_at:
        type: string
      foto:
        type: string
      history:
        items:
          $ref: '#/definitions/models.HistoryEntry'
        type: array
      hp:
        type: string
      id:
        type: string
      id_mahasiswa:
        type: string
      id_status_mahasiswa:
        type: string
      jenis_jalur_mahasiswa_key:
        type: integer
      jk:
        type: string
      jurusan_key:
        type: integer
      kelurahan:
        type: string
      keterangan:
        type: string
      kewarganegaraan:
        type: string
      kode_kabupaten:
        type: string
      kode_kecamatan:
        type: string
      kode_pos:
        type: string
      kode_provinsi:
        type: string
      kurikulum:
        type: string
      last_status:
        type: string
      metadata_key:
        type: integer
      nama:
        type: string
      nama_gadis_ibu_kandung:
        type: string
      nik:
        type: string
      nim:
        type: string
      nisn:
        type: string
      no_kps:
        type: string
      npsn_sekolah:
        type: string
      penerima_bidikmisi:
        type: string
      permalink:
        type: string
      rt:
        type: string
      rw:
        type: string
      semester_posisi:
        type: integer
      source_key:
        type: string
      spp:
        type: string
      status:
        type: string
      sumber:
        type: string
      tahun_masuk:
        type: string
      telp:
        type: string
      tempatkuliah:
        type: string
      tgl_lahir:
        type: string
      tgl_mulai_kuliah:
        type: string
      tmp_lahir:
        type: string
      uk_id:
        type: string
      uk_kode:
        type: string
      uk_program:
        type: string
      unit:
        $ref: '#/definitions/models.Unit'
      updated_at:
        type: string
      validasi:
        type: string
      wa:
        type: string
    type: object
  models.MahasiswaHistoryResponse:
    properties:
      email_sso:
        type: string
      fakultas:
        type: string
      jurusan:
        type: string
      kewarganegaraan:
        type: string
      nama:
        type: string
      nama_pa:
        type: string
      nim:
        type: string
      periode:
        type: string
      prodi:
        type: string
      semester:
        type: integer
      status:
        type: string
      tahun:
        type: integer
      tahun_masuk:
        type: string
      telp:
        type: string
    type: object
...
  models.PaginationResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      pages:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.Unit:
    properties:
      fakultas:
        type: string
      fkt_kode:
        type: string
      jrs_kode:
        type: string
      jurusan:
        type: string
      prd_kode:
        type: string
      prodi:
        type: string
      uk_kode:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: anggadek857@gmail.com
    name: Dwi Angga
    url: https://github.com/anggananda
  description: |-
    This repository contains the **RESTful API** for the **Executive Information System (EIS) Undiksha**, built with **Golang** using the **Gin framework**.
    The API consumes data from **MongoDB** (populated via ETL from SQL in a separate repository) and uses **Redis** for caching.
    The system is integrated with **Undiksha Single Sign-On (SSO)** and implements **JWT Authentication** for secure access.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/anggananda
  title: Executive Information System RESTful API
  version: "1.0"
paths:
...
  /kritik-saran:
    get:
      consumes:
      - application/json
      description: Mendapatkan data kritik & saran berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.KritikSaran'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered kritik & saran
      tags:
      - Kritik & Saran
  /kritik-saran/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar kritik saran yang telah difilter ke dalam file
        CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Kritik Saran ke CSV
      tags:
      - Kritik & Saran
...
  /mhs/{nim}:
    get:
      consumes:
      - application/json
      description: Mendapatkan detail data mahasiswa
      parameters:
      - description: nim
        in: query
        name: nim
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListDetailResponse'
            - properties:
                datas:
                  $ref: '#/definitions/models.Mahasiswa'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get detail mahasiswa
      tags:
      - Mahasiswa
  /mhs/history:
    get:
      consumes:
      - application/json
      description: Mendapatkan data mahasiswa berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: integer
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: integer
      - description: Filter berdasarkan angkatan
        in: query
        name: angkatan
        type: integer
      - description: Filter berdasarkan status
        in: query
        name: status
        type: integer
      - description: Filter berdasarkan kewarganegaraan
        in: query
        name: kewarganegaraan
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.MahasiswaHistoryResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered mahasiswa
      tags:
      - Mahasiswa
  /mhs/history/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar mahasiswa yang telah difilter ke dalam file CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: integer
      - description: Filter berdasarkan angkatan
        in: query
        name: angkatan
        type: integer
      - description: Filter berdasarkan status
        in: query
        name: status
        type: integer
      - description: Filter berdasarkan kewarganegaraan
        in: query
        name: kewarganegaraan
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Mahasiswa ke CSV
      tags:
      - Mahasiswa
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"